Move player "shouldUpdate" and UPDATE_COOLDOWN
